Update files as part of modularization
