+++ invokeai/backend/stable_diffusion/onnx_pipeline.py
@@ -20,7 +20,7 @@ class txt2img:
         self.num_images_per_prompt = num_images
         self.num_inference_steps = steps
 
-    def onnx_txt2img(self, prompt, model, precision):
+    def onnx_txt2img(self, prompt, model, precision, outdir):
         #model ="runwayml/stable-diffusion-v1-5" or "CompVis/stable-diffusion-v1-4"
         if precision == "cpu":
             onnx_pipe = OnnxStableDiffusionPipeline.from_pretrained(model, revision="onnx", provider="CPUExecutionProvider")
@@ -30,7 +30,7 @@ class txt2img:
         try:
             image = onnx_pipe(prompt, self.height, self.height, num_images_per_prompt=self.num_images_per_prompt, num_inference_steps=self.num_inference_steps).images[0]
             timestamp = int(time.time())
-            image.save(f"Inference_{timestamp}.png")
+            image.save(f"./{outdir}/Inference_{timestamp}.png")
 
         except (FileNotFoundError, TypeError, AssertionError) as e:
             print("Execption: ", e)
